Add files via upload
as mentioned in the gc, these files will not work out-of-the-box as they've been using hypothetical variables as a stand-in. i have NOT had the ability to test them because of this issue - regard highly with errors. ConnectionEngine is also pretty "unfinished" considering a mishap within groupwork that led to one teammate essentially completing half the work of another